fix: review adjustments
